:bricks: Load balancer setup
Cloud Run doesn't seem to support mapping domains from CloudFlare CNAME.
Introduce a load balancer for more flexibility.
On the CloudFlare side we can simply point a A DNS type to the LB IP.
Note this setup is optional and disabled by default, but can be enabled
via the `create_load_balancer` variable.